Separating the blueprints all into their own modules.
